requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.13.0
kornia
tqdm
pandas
scipy
accelerate
deepspeed
torchsnooper
matplotlib
pycocotools
h5py
nibabel
urllib3
numpy
wandb
pillow
scikit-image
clip
clip-retrieval
transformers
gpustat